Versao final 5.0

+++ Rede P2P/src/br/com/ufes/aplicacaoP2P/Cliente.java
@@ -57,7 +57,7 @@ public class Cliente {
 		
 		//Aloca um espaço de 5 bytes
 		ByteBuffer sendData = ByteBuffer.allocate(21);
-		byte codeMessage[] = {1};  //Codigo da Mensagem Leave
+		byte codeMessage[] = {(byte) 1};  //Codigo da Mensagem Leave
 		sendData.put(codeMessage);
 		sendData.put(intToBytes(idNodeLeaving));
 		sendData.put(intToBytes(idSuc));
